rename local train to local baseline model
