+++ federatedml/conf/setting_conf/LocalBaselinModel.json
@@ -1,0 +1,12 @@
+{
+    "module_path":  "federatedml/local_baseline_model/",
+    "default_runtime_conf": "local_baseline_model_param.json",
+    "param_class" : "federatedml/param/local_baseline_model_param.py/LocalBaselineModelParam",
+    "role":
+    {
+        "guest":
+        {
+            "program": "local_train.py/LocalBaselineModel"
+        }
+    }
+}
